+++ DEVELOPMENT.md
@@ -5,11 +5,10 @@
 
 It is suggested you use [nvm](https://github.com/nvm-sh/nvm) to manage your node version
 
-
 It is suggested to install this globally[@xarc/run-cli](https://www.npmjs.com/package/@xarc/run-cli)
 
-This will allow you to run varios tasks 
-![xrun auto complete](./files/xrun-autocomplete.png)
+This will allow you to run varios tasks
+![xrun auto complete](./screenshots/xrun-autocomplete.png)
 
 In devmode you can open [local remote devtools](http://localhost:8000)
-![remote redux devtools](./files/remote-redux-devtools.png)
+![remote redux devtools](./screenshots/remote-redux-devtools.png)
